requirements.base.txt
biopython @ git+https://github.com/netogallo/biopython.git@a95ae4130580d09107882ee6bdbc159d4803b122
docopt
igraph
open3d
openpyxl
pandas
primer3-py
pydantic
pyinstaller
requests
SQLAlchemy
